pyproject.toml
[tool.poetry]
name = "a-qsi-api-v3-client"
version = "v3"
description = "A client library for accessing aQsi API V3"
authors = []
readme = "README.md"
packages = [
    {include = "a_qsi_api_v3_client"},
]
include = ["CHANGELOG.md", "a_qsi_api_v3_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
